+++ src/db/traits.rs
@@ -9,6 +9,7 @@ pub trait Db {
     fn week_tasks(&self, week: &u32) -> Vec<Task>;
     fn active_task(&self) -> Result<Task, Error>;
     fn task(&self, id: &i64) -> Result<Task, Error>;
+    fn last_task(&self) -> Result<Task, Error>;
     fn stop_task(&self, id: &i64) -> Result<Task, Error>;
     fn add_task(&self, project: &String, desc: &String, external_id: &Option<String>) -> Result<(), Error>;
     fn change_task_desc(&self, id: &i64, desc: &String) -> Result<(), Error>;